Replaced libao by our own plugin system. At the moment, there's only an OSS plugin, ALSA will probably follow soon. See ChangeLog for details ;)
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@7691 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ misc/eplayer/src/playlist.h
@@ -17,7 +17,7 @@ typedef struct {
 	char artist[PLAYLIST_ITEM_COMMENT_LEN];
 	char title[PLAYLIST_ITEM_COMMENT_LEN];
 	char album[PLAYLIST_ITEM_COMMENT_LEN];
-	double duration;
+	int duration;
 
 	int channels; /* number of channels */
 	long rate; /* bitrate */
